Extract remaining CLI command logic into cli_api.
Finish Phase-1 library-first extraction so setup, info, edit, pull, new, and
serve are callable without a Typer context; thin cli.py adapters pass
typer.echo. Also fix the cookiecutter ModuleNotFoundError continue-bug.
