+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Complete cli_api extraction for remaining CLI commands (new, serve, setup, …) (#651).
+
 * Fix summary CV-split plots: use exclude_step_types and full−non_cv instead of dead selector_type (#654).
 
 * Wire Batch.plot to plotting and delete batch_plotters.py (#658).
